use simd_json in debezium/maxwell json parser & fix bugs
