+++ e2e_test/source/kafka.slt
@@ -167,6 +167,32 @@ WITH (
     scan.startup.mode = 'earliest')
 ROW format JSON;
 
+# we cannot create debezium source without pk
+statement error
+create materialized source s13 (
+  id integer,
+  name varchar,
+  is_adult integer,
+  birthday timestamp
+) with (
+  connector = 'kafka',
+  topic = 'maxwell_json',
+  properties.bootstrap.server = '127.0.0.1:29092'
+) row format maxwell;
+
+statement ok
+create materialized source s13 (
+  id integer,
+  name varchar,
+  is_adult integer,
+  birthday timestamp,
+  PRIMARY KEY (id)
+) with (
+  connector = 'kafka',
+  topic = 'maxwell_json',
+  properties.bootstrap.server = '127.0.0.1:29092'
+) row format maxwell;
+
 statement ok
 flush;
 
@@ -312,6 +338,12 @@ select id, code, timestamp, xfas, contacts from s12;
 ----
 100 abc 1473305798 {"(0,200,10.0.0.1)","(1,400,10.0.0.2)"} ({1xxx,2xxx},{1xxx,2xxx})
 
+query ITIT
+select * from s13 order by id;
+----
+1 tom 0 2017-12-31 16:00:01
+2 chi 1 1999-12-31 16:00:01
+
 statement ok
 drop materialized view source_mv1
 
@@ -335,3 +367,6 @@ drop source s11
 
 statement ok
 drop source s12
+
+statement ok
+drop source s13
